Prepare release of first working version v0.0.1
Provide the 'Heatseeker: Apply' command to color the lines in the active
editor by querying the server in the 'heatseeker.host' configuration for
statistics on previous trace results. Switching the active editor will
automatically apply color to the new file, but statistics are cached.

Also, add an initial icon and at least a bit of information the README.